backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.6
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   langchain
    #   starlette
    #   watchfiles
asttokens==2.4.1
    # via icontract
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
beartype==0.16.4
    # via perscache
black==23.10.1
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.2
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   black
    #   pip-tools
    #   uvicorn
cloudpickle==3.0.0
    # via perscache
dataclasses-json==0.6.1
    # via langchain
diskcache==5.6.3
    # via pyautogen
dnspython==2.4.2
    # via email-validator
email-validator==2.1.0.post1
    # via pydantic
exceptiongroup==1.1.3
    # via anyio
fastapi==0.104.0
    # via -r requirements.in
filelock==3.13.1
    # via torch
flake8==6.1.0
    # via -r requirements.in
flaml==2.1.1
    # via pyautogen
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.10.0
    # via torch
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
icontract==2.6.4
    # via perscache
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via
    #   build
    #   numba
interegular==0.3.2
    # via outlines
jinja2==3.1.2
    # via
    #   outlines
    #   torch
joblib==1.3.2
    # via outlines
jsonpatch==1.33
    # via langchain
jsonpointer==2.4
    # via jsonpatch
langchain==0.0.330
    # via -r requirements.in
langsmith==0.0.57
    # via langchain
lark==1.1.8
    # via outlines
llvmlite==0.41.1
    # via numba
markupsafe==2.1.3
    # via jinja2
marshmallow==3.20.1
    # via dataclasses-json
mccabe==0.7.0
    # via flake8
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
networkx==3.1
    # via torch
numba==0.58.1
    # via outlines
numpy==1.24.4
    # via
    #   flaml
    #   langchain
    #   numba
    #   outlines
    #   scipy
openai==0.28.1
    # via
    #   -r requirements.in
    #   pyautogen
outlines==0.0.9
    # via -r requirements.in
packaging==23.2
    # via
    #   black
    #   build
    #   marshmallow
pathspec==0.11.2
    # via black
perscache==0.6.1
    # via outlines
pillow==10.1.0
    # via outlines
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.11.0
    # via black
pyautogen==0.1.14
    # via -r requirements.in
pycodestyle==2.11.1
    # via flake8
pydantic[email]==2.4.2
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langsmith
    #   outlines
    #   pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-settings==2.0.3
    # via -r requirements.in
pyflakes==3.1.0
    # via flake8
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via
    #   pyautogen
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.1
    # via
    #   langchain
    #   uvicorn
requests==2.31.0
    # via
    #   langchain
    #   langsmith
    #   openai
scipy==1.10.1
    # via outlines
six==1.16.0
    # via asttokens
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.23
    # via langchain
starlette==0.27.0
    # via fastapi
sympy==1.12
    # via torch
tenacity==8.2.3
    # via
    #   langchain
    #   outlines
termcolor==2.3.0
    # via pyautogen
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-hooks
torch==2.1.0
    # via outlines
tqdm==4.66.1
    # via openai
typing-extensions==4.8.0
    # via
    #   annotated-types
    #   black
    #   fastapi
    #   icontract
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   torch
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.0.7
    # via requests
uvicorn[standard]==0.23.2
    # via -r requirements.in
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
wheel==0.41.2
    # via pip-tools
yarl==1.9.2
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
